DualBounds-Example/requirements.txt
python==3.7.7
argparse==1.4.0
gurobipy==10.0.2
matplotlib==3.5.3
networkx==2.5
geopandas==0.10.2
scipy==1.7.3
numpy==1.20.1
